+++ test/configuration_test.rb
@@ -10,7 +10,7 @@ class ConfigurationTest < Test::Unit::TestCase
       assert_config_default :compression_schema,   [6,5,3,4,4,3]
       assert_config_default :frequency,            10
       assert_config_default :maximum_datapoints,   800
-      assert_config_default :data_type,            :float
+      assert_config_default :data_type,            :double
       assert_config_default :database_config_file, File.expand_path('../../config/database.yml', __FILE__)
     end
